% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Hello World function}
\usage{
hello(x)
}
\arguments{
\item{x}{some name}
}
\value{
The output from \code{\link{print}}
}
\description{
This function simply prints \strong{Hello World}
}
\examples{
hello("Shashank")
\dontrun{
hello("shashank")
}
}
