% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printsum}
\alias{printsum}
\title{printsum function}
\usage{
printsum(x, y)
}
\arguments{
\item{x}{number}

\item{y}{number}
}
\value{
returns sum of x and y
}
\description{
This function simply prints sum of x and y
}
\examples{
printsum(1, 2)
\dontrun{
printsum(4, 5)
}
}
